# openapi.yaml
openapi: 3.0.3
info:
  title: LifeBalance Sprint 3 Resource Capital and Task Core API
  version: 1.0.0
  description: >
    OpenAPI contract for LB-828 covering Sprint 3 Resource Capital Foundation
    and Task Core endpoints. All documented responses use the LifeBalance
    ApiResponse envelope: success, data, error, and timestamp.
  contact:
    name: LifeBalance Team
    email: team@lifebalance.com
  license:
    name: LifeBalance
    url: https://lifebalance.com/license
servers:
  - url: /
    description: Current service or gateway host
security:
  - bearerAuth: []
...
paths:
  /api/v1/capital-cycles:
    get:
      tags:
        - Capital Cycles
      summary: List capital cycles
      description: List authenticated user's capital cycles with optional type, status, period, pagination, and sorting filters.
      operationId: listCapitalCycles
      parameters:
        - $ref: '#/components/parameters/CapitalCycleTypeQuery'
        - $ref: '#/components/parameters/CapitalCycleStatusQuery'
        - name: fromDate
          in: query
          required: false
          description: Include cycles ending on or after this date.
          schema:
            type: string
            format: date
          example: '2026-08-01'
        - name: toDate
          in: query
          required: false
          description: Include cycles starting before or on this date.
          schema:
            type: string
            format: date
          example: '2026-08-31'
        - $ref: '#/components/parameters/PageQuery'
        - $ref: '#/components/parameters/SizeQuery'
        - $ref: '#/components/parameters/SortQuery'
      responses:
        '200':
          description: Capital cycles returned successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseCapitalCyclePage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags:
        - Capital Cycles
      summary: Create capital cycle
      description: Create a new user-owned capital cycle in DRAFT status after validating period rules and overlap constraints.
      operationId: createCapitalCycle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCapitalCycleRequest'
            example:
              name: August 2026 monthly cycle
              description: Main planning cycle for August 2026.
              type: MONTHLY
              startDate: '2026-08-01'
              endDate: '2026-08-31'
      responses:
        '201':
          description: Capital cycle created successfully.
          headers:
            Location:
              description: URL of the created capital cycle.
              schema:
                type: string
              example: /api/v1/capital-cycles/22222222-2222-2222-2222-222222222222
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseCapitalCycle'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/capital-cycles/active:
    get:
      tags:
        - Capital Cycles
      summary: Get active capital cycle
      description: Get the authenticated user's active capital cycle, optionally filtered by cycle type.
      operationId: getActiveCapitalCycle
      parameters:
        - $ref: '#/components/parameters/CapitalCycleTypeQuery'
      responses:
        '200':
          description: Active capital cycle returned successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseCapitalCycle'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/capital-cycles/{id}:
    get:
      tags:
        - Capital Cycles
      summary: Get capital cycle
      description: Get an authenticated user's capital cycle detail by id.
      operationId: getCapitalCycle
      parameters:
        - $ref: '#/components/parameters/CapitalCycleIdPath'
      responses:
        '200':
          description: Capital cycle returned successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseCapitalCycle'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags:
        - Capital Cycles
      summary: Update capital cycle
      description: Update editable information for an authenticated user's DRAFT capital cycle.
      operationId: updateCapitalCycle
      parameters:
        - $ref: '#/components/parameters/CapitalCycleIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCapitalCycleRequest'
            example:
              name: August 2026 focus cycle
              description: Updated planning scope for August 2026.
              type: MONTHLY
              startDate: '2026-08-01'
              endDate: '2026-08-31'
      responses:
        '200':
          description: Capital cycle updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseCapitalCycle'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags:
        - Capital Cycles
      summary: Delete capital cycle
      description: Delete an authenticated user's DRAFT capital cycle only when no dependent capital records exist.
      operationId: deleteCapitalCycle
      parameters:
        - $ref: '#/components/parameters/CapitalCycleIdPath'
      responses:
        '204':
          description: Capital cycle deleted successfully.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/capital-cycles/{id}/activate:
    post:
      tags:
        - Capital Cycles
      summary: Activate capital cycle
      description: Activate a DRAFT or REOPENED capital cycle after ownership, status transition, and one-active-cycle validation.
      operationId: activateCapitalCycle
      parameters:
        - $ref: '#/components/parameters/CapitalCycleIdPath'
      responses:
        '200':
          description: Capital cycle activated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseCapitalCycle'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/capital-cycles/{id}/close:
    post:
      tags:
        - Capital Cycles
      summary: Close capital cycle
      description: Close an ACTIVE or REOPENED capital cycle with a required business reason.
      operationId: closeCapitalCycle
      parameters:
        - $ref: '#/components/parameters/CapitalCycleIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CloseCapitalCycleRequest'
            example:
              reason: Monthly cycle completed.
      responses:
        '200':
          description: Capital cycle closed successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseCapitalCycle'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/capital-cycles/{id}/reopen:
    post:
      tags:
        - Capital Cycles
      summary: Reopen capital cycle
      description: Reopen a CLOSED capital cycle with a required business reason.
      operationId: reopenCapitalCycle
      parameters:
        - $ref: '#/components/parameters/CapitalCycleIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReopenCapitalCycleRequest'
            example:
              reason: Need correction after late transaction import.
      responses:
        '200':
          description: Capital cycle reopened successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseCapitalCycle'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Keycloak access token using the Bearer authentication scheme.
  parameters:
    CapitalCycleIdPath:
      name: id
      in: path
      required: true
      description: Capital cycle id.
      schema:
        type: string
        format: uuid
      example: 22222222-2222-2222-2222-222222222222
...
    CapitalCycleTypeQuery:
      name: type
      in: query
      required: false
      description: Filter by capital cycle type.
      schema:
        $ref: '#/components/schemas/CapitalCycleType'
      example: MONTHLY
    CapitalCycleStatusQuery:
      name: status
      in: query
      required: false
      description: Filter by capital cycle status.
      schema:
        $ref: '#/components/schemas/CapitalCycleStatus'
      example: ACTIVE
    PageQuery:
      name: page
      in: query
      required: false
      description: Zero-based page index.
      schema:
        type: integer
        format: int32
        minimum: 0
        default: 0
      example: 0
    SizeQuery:
      name: size
      in: query
      required: false
      description: Page size.
      schema:
        type: integer
        format: int32
        minimum: 1
        maximum: 200
        default: 20
      example: 20
    SortQuery:
      name: sort
      in: query
      required: false
      description: Sort expression in Spring Pageable format, for example createdAt,desc.
      schema:
        type: array
        items:
          type: string
      style: form
      explode: true
      example:
        - createdAt,desc
        - id,desc
  responses:
    BadRequest:
      description: Bad Request or validation error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponseError'
          example:
            success: false
            data: null
            error:
              code: COMMON_VALIDATION_FAILED
              message: Request validation failed
              details:
                name: must not be blank
            timestamp: '2026-08-14T00:00:00Z'
    Unauthorized:
      description: Authentication is missing or invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponseError'
          example:
            success: false
            data: null
            error:
              code: AUTH_UNAUTHORIZED
              message: Authentication is required
              details: {}
            timestamp: '2026-08-14T00:00:00Z'
    Forbidden:
      description: Authenticated user does not have permission to access the resource.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponseError'
          example:
            success: false
            data: null
            error:
              code: AUTH_FORBIDDEN
              message: Access is denied
              details: {}
            timestamp: '2026-08-14T00:00:00Z'
    NotFound:
      description: Requested resource was not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponseError'
          example:
            success: false
            data: null
            error:
              code: COMMON_NOT_FOUND
              message: Resource was not found
              details: {}
            timestamp: '2026-08-14T00:00:00Z'
    Conflict:
      description: Business conflict.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponseError'
          example:
            success: false
            data: null
            error:
              code: RESOURCE_CONFLICT
              message: Business rule conflict
              details: {}
            timestamp: '2026-08-14T00:00:00Z'
    InternalServerError:
      description: Internal server error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponseError'
          example:
            success: false
            data: null
            error:
              code: COMMON_INTERNAL_ERROR
              message: Unexpected server error
              details: {}
            timestamp: '2026-08-14T00:00:00Z'
  schemas:
    ApiResponseBase:
      type: object
      required:
        - success
        - timestamp
      properties:
        success:
          type: boolean
          example: true
        timestamp:
          type: string
          format: date-time
          example: '2026-08-14T00:00:00Z'
    ApiResponseError:
      allOf:
        - $ref: '#/components/schemas/ApiResponseBase'
        - type: object
          required:
            - success
            - error
            - timestamp
          properties:
            success:
              type: boolean
              example: false
            data:
              type: object
              nullable: true
              description: Always null for failed responses.
            error:
              $ref: '#/components/schemas/ApiError'
...
    ApiResponseCapitalCycle:
      allOf:
        - $ref: '#/components/schemas/ApiResponseBase'
        - type: object
          required:
            - data
          properties:
            data:
              $ref: '#/components/schemas/CapitalCycleResponse'
            error:
              type: object
              nullable: true
              allOf:
                - $ref: '#/components/schemas/ApiError'
    ApiResponseCapitalCyclePage:
      allOf:
        - $ref: '#/components/schemas/ApiResponseBase'
        - type: object
          required:
            - data
          properties:
            data:
              $ref: '#/components/schemas/PageCapitalCycleResponse'
            error:
              type: object
              nullable: true
              allOf:
                - $ref: '#/components/schemas/ApiError'
...
    ApiError:
      type: object
      required:
        - code
        - message
        - details
      properties:
        code:
          type: string
          example: COMMON_VALIDATION_FAILED
        message:
          type: string
          example: Request validation failed
        details:
          type: object
          additionalProperties:
            type: string
          example:
            name: must not be blank
    CreateCapitalCycleRequest:
      type: object
      required:
        - name
        - type
        - startDate
        - endDate
      properties:
        name:
          type: string
          maxLength: 255
          example: August 2026 monthly cycle
        description:
          type: string
          nullable: true
          maxLength: 2000
          example: Main planning cycle for August 2026.
        type:
          $ref: '#/components/schemas/CapitalCycleType'
        startDate:
          type: string
          format: date
          example: '2026-08-01'
        endDate:
          type: string
          format: date
          example: '2026-08-31'
    UpdateCapitalCycleRequest:
      allOf:
        - $ref: '#/components/schemas/CreateCapitalCycleRequest'
      description: Same payload as create. startDate must not be after endDate, and updates are allowed only for DRAFT cycles.
    CloseCapitalCycleRequest:
      type: object
      required:
        - reason
      properties:
        reason:
          type: string
          maxLength: 1000
          example: Monthly cycle completed.
    ReopenCapitalCycleRequest:
      type: object
      required:
        - reason
      properties:
        reason:
          type: string
          maxLength: 1000
          example: Need correction after late transaction import.
...
    CapitalCycleResponse:
      type: object
      required:
        - id
        - name
        - type
        - startDate
        - endDate
        - status
        - overAllocationAllowed
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
          example: 22222222-2222-2222-2222-222222222222
        name:
          type: string
          example: August 2026 monthly cycle
        description:
          type: string
          nullable: true
          example: Main planning cycle for August 2026.
        type:
          $ref: '#/components/schemas/CapitalCycleType'
        startDate:
          type: string
          format: date
          example: '2026-08-01'
        endDate:
          type: string
          format: date
          example: '2026-08-31'
        status:
          $ref: '#/components/schemas/CapitalCycleStatus'
        overAllocationAllowed:
          type: boolean
          example: false
        activatedAt:
          type: string
          format: date-time
          nullable: true
          example: '2026-08-01T00:00:00Z'
        closedAt:
          type: string
          format: date-time
          nullable: true
          example: null
        reopenedAt:
          type: string
          format: date-time
          nullable: true
          example: null
        closeReason:
          type: string
          nullable: true
          example: null
        reopenReason:
          type: string
          nullable: true
          example: null
        createdAt:
          type: string
          format: date-time
          example: '2026-08-01T00:00:00Z'
        updatedAt:
          type: string
          format: date-time
          example: '2026-08-01T00:00:00Z'
...
    PageCapitalCycleResponse:
      allOf:
        - $ref: '#/components/schemas/PageBase'
        - type: object
          required:
            - content
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/CapitalCycleResponse'
...
    CapitalCycleType:
      type: string
      enum:
        - DAILY
        - WEEKLY
        - MONTHLY
      example: MONTHLY
    CapitalCycleStatus:
      type: string
      enum:
        - DRAFT
        - ACTIVE
        - CLOSED
        - REOPENED
      example: ACTIVE
...